Refactor: Separate styles from its component
