+++ flight-reservation/components/CheckBoxComponent/styles.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+    container: {
+      marginVertical: 10,
+    },
+    innerIconStyle: {
+      borderWidth: 2,
+    },
+    fillColor: 'gray'
+    ,
+    size: 25
+    ,
+    checkedFillColor: 'blue',
+  });
+
