+++ flight-reservation/components/index.js
@@ -2,4 +2,4 @@ import TextInputComponent from './TextInputComponent';
 import ButtonComponent from './ButtonComponent'
 import CheckBoxComponent from './CheckBoxComponent'
 
-export {TextInputComponent, ButtonComponent, CheckBoxComponent }
+export {TextInputComponent, ButtonComponent, CheckBoxComponent}
