+++ flight-reservation/screens/SignUpScreen/index.js
@@ -1,4 +1,5 @@
-import { StyleSheet, Text, TextInput, View, Platform, StatusBar } from "react-native";
+import {  Text , View  } from "react-native";
+import { styles } from "./styles";
 import React from "react";
 import { TextInputComponent, ButtonComponent, CheckBoxComponent } from "../../components";
 
@@ -8,7 +9,7 @@ const SignUpScreen = () => {
       <Text style={styles.title}>Sign Up</Text>
       <TextInputComponent title="First name"/>
       <TextInputComponent title="Email *"/>
-      <TextInputComponent title="Password *" hint={'Use 8 or more characters with a mix of letters, numbers, and symbols.'}/>
+      <TextInputComponent title="Password *" help={'Use 8 or more characters with a mix of letters, numbers, and symbols.'}/>
 
       <CheckBoxComponent text='I agree to the Terms and Privacy Policy'/>
       <CheckBoxComponent text='Subscribe for select product updates.'/>
@@ -19,8 +20,8 @@ const SignUpScreen = () => {
       </View>
       <ButtonComponent title='Sign Up with Google' textColor='#fff' bgColor='gray' icon="google"/>
       <View style={styles.loginSection}>
-        <Text style={{color:'grey'}}>Already have an account?</Text>
-        <Text style={{color:'blue',textDecorationLine:'underline'}}>Log In</Text>
+        <Text style={styles.textColor}>Already have an account?</Text>
+        <Text style={styles.linkColor}>Log In</Text>
       </View>
 
     </View>
@@ -28,32 +29,3 @@ const SignUpScreen = () => {
 };
 
 export default SignUpScreen;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    width: '90%',
-    marginHorizontal: 'auto',
-  },
-  inputContainer: {
-
-  },
-  title: {
-    color: '#6270de',
-    fontWeight: '800',
-    fontSize:24,
-    paddingVertical:30,
-  },
-  orSection: {
-    alignItems: 'center',
-    marginVertical:10,
-  },
-  loginSection:{
-    flexDirection:'row',
-    justifyContent:'center',
-    paddingTop:20,
-    paddingBottom:10,
-  },
- 
-  
-});
